+++ dist-cjs/src/cli/commander-fix.js
@@ -1,5 +1,0 @@
-import { Command as CommandConstructor } from 'commander';
-export const Command = CommandConstructor;
-export default Command;
-
-//# sourceMappingURL=commander-fix.js.map