+++ dist-cjs/src/cli/commands/hook-types.js
@@ -1,3 +1,0 @@
-export { };
-
-//# sourceMappingURL=hook-types.js.map