+++ dist-cjs/src/cli/commands/start.js
@@ -1,3 +1,0 @@
-export { startCommand } from './start/index.js';
-
-//# sourceMappingURL=start.js.map