+++ dist-cjs/src/cli/simple-commands/init.js
@@ -1,3 +1,0 @@
-export { initCommand } from './init/index.js';
-
-//# sourceMappingURL=init.js.map