+++ dist-cjs/src/cli/simple-commands/start.js
@@ -1,3 +1,0 @@
-export { startCommand } from './start-wrapper.js';
-
-//# sourceMappingURL=start.js.map