+++ dist-cjs/src/mcp/transports/base.js
@@ -1,3 +1,0 @@
-export { };
-
-//# sourceMappingURL=base.js.map