+++ dist-cjs/src/types/cliffy.d.js
@@ -1,2 +1,0 @@
-
-//# sourceMappingURL=cliffy.d.js.map