+++ dist-cjs/src/types/global.d.js
@@ -1,3 +1,0 @@
-export { };
-
-//# sourceMappingURL=global.d.js.map